requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pypdf
requests
openai>=1.0.0